Validate and apply addons as phases
